+++ Application/Commands/Users/FollowUser/FollowUserCommandHandler.cs
@@ -1,34 +1,30 @@
-using Application.Interfaces;
-using AutoMapper;
 using MediatR;
 using Domain.Entities;
+using Application.Interfaces.Repositories;
 
 namespace Application.Commands.Users.FollowUser
 {
-    public class FollowUserCommandHandler : IRequestHandler<FollowUserCommand>
+    public class FollowUserCommandHandler(IUserRepository userRepository) : IRequestHandler<FollowUserCommand>
     {
-        private readonly IUserRepository _userRepository;
-        private readonly IMapper _mapper;
-
-        public FollowUserCommandHandler(IUserRepository userRepository, IMapper mapper)
-        {
-            _userRepository = userRepository;
-            _mapper = mapper;
-        }
+        private readonly IUserRepository _userRepository = userRepository;
 
         public async Task<Unit> Handle(FollowUserCommand request, CancellationToken cancellationToken)
-        {   
-            var user = _userRepository.GetByIdAsync(request.UserId).Result;
-            var followUser = _userRepository.GetByIdAsync(request.FollowUserId).Result;
+        {
+            var user = await _userRepository.GetByIdAsync(request.UserId);
+            var followUser = await _userRepository.GetByIdAsync(request.FollowUserId);
+
+            if (user == null || followUser == null || user.FollowedUsers == null)
+                return Unit.Value;
 
             var isFollowed = user.FollowedUsers.FirstOrDefault(x => x.FollowedUserId == followUser.Id);
+
             if (isFollowed != null)
             {
                 await _userRepository.DeleteFollowerAsync(isFollowed);
             }
             else
             {
-                var follower = new Follower 
+                var follower = new Follower
                 {
                     UserId = user.Id,
                     User = user,
@@ -37,7 +33,7 @@ namespace Application.Commands.Users.FollowUser
                 };
                 await _userRepository.AddFollowerAsync(follower);
             }
-       
+
             return Unit.Value;
         }
     }
